Check if lock is owned before unlocking

+++ core/src/com/unciv/ui/worldscreen/mainmenu/DropBox.kt
@@ -267,6 +267,9 @@ class ServerMutex(val gameInfo: GameInfoPreview) {
      * @see lock
      */
     fun unlock() {
+        if (!locked)
+            return
+
         DropBox.deleteFile("${OnlineMultiplayer().getGameLocation(gameInfo.gameId)}_Lock")
         locked = false
     }
